requirements.txt
python-bale-bot>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
googlesearch-python>=1.2.3
twikit>=2.3.0
httpx>=0.27.0
pillow>=10.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
selenium>=4.0.0
webdriver-manager>=4.0.0
